+++ libraries/base/changelog.md
@@ -97,6 +97,9 @@
     are swapped, such that `Data.List.nubBy (<) [1,2]` now returns `[1]`
     instead of `[1,2]` (#2528, #3280, #7913)
 
+  * New module `Data.Functor.Identity` (previously provided by `transformers`
+    package). (#9664)
+
 ## 4.7.0.1  *Jul 2014*
 
   * Bundled with GHC 7.8.3